New method to cleanup all data associated with a community in the throttle object:
if a community is deleted when a user is accessing it, some leftover information will be kept in the object. At the next request this will be interpreted as the community having been left and the system will try to track this event into request_monitor_community_activities. Depending on who comes first, we might try to insert a tuple in the table for an object that does not exist anymore.
Hence, we use this method to cleanup all community info before the actual deletion in dotlrn (happening only downstream).